Experiment with kotlin-poet without builders
I thought the builder pattern was mostly useless in kotlin,
which has named parameters, default values and works well
with an immutable functional style like the one used by kotlin poet.

This is an experiment with the kotlinpoet codebase to test this hypothesis.

Of course I don't advocate removing the builders apis,
that would be bad for people calling kotlinpoet from java or with an existing code base.
I'm only adding stuff, and delegating to those builders,
even adding a configuration lambda for all the corner cases where a builder can be useful

I modified the examples in the README to see what the changes would look like