Proposal - Introduce %C for strings that should be compile-time constants
This is a proposal solution for addressing strings that should be compile-time constants (such as for use in annotation members or `const` declarations). Currently, KotlinPoet will emit strings with newlines as templated strings with `trimMargin()`

Will also open a separate change with a plumbing approach for comparison. I wasn't sure which would be better, as plumbing seemed to require pushing state into `CodeBlock` itself