Introduce notion of "constant context" to CodeWriter (#654)
* Add isConstantContext to stringLiteralWithQuotes

* Add const property test

* Add annotationStringsAreConstant test

* Add literal annotations tests

* Add isConstantContext parameter to CodeWriter#emitCode()

This can be used as signal for multiple conditions to know if they're emitting into a context that should be considered constant, such as annotation parameters or const properties

* Wire isConstantContext into PropertySpec

only gate on CONST presence. If developers try to define types that aren't supported as const, that's on them!

* Wire in isConstantContext for annotation spec member emission

This flips the previous default, where all literal annotations were without the @ sign. Now this clarifies that to only be within the context of another annotation's members

* Use trimMargin