Add experimental typeNameOf() API for TypeName
This depends on the experimental `typeOf()` API from the stdlib and propagates the same experimental annotation to guard it.

This can be used as a shorthand for basically almost all of the current typename variants' static factories, and just leave the factories for when arguments are being passed in from somewhere else.