Add support for detecting Kotlin default interface functions (#771)
* Add support for detecting Kotlin default interface functions

Will still need to add more tests, but putting up for early review.

This adds support for detecting interface functions in Kotlin that have default implementations. This probably merits a separate deep dive blogpost, but in short they way they are implement is a synthetic `DefaultImpls` class in the interface with static functions containing the default implementation. The rules are pretty simple:

- method name is the same
- parameters are the same
- the target interface instance is added as a parameter to the beginning of these parameters

So by looking for the existing of that `DefaultImpls` class + method with the modified signature, we can infer if there is a default implementation for Kotlin. There are slightly different rules for when default parameter arguments are in play, but these methods still exist in those cases too

In short:

```kotlin
interface Foo {
  fun bar(input: String) {
    println("I'm a " + this + " saying " + input)
  }
}
```

Becomes roughly this in bytecode:

```java
interface Foo {

  void bar();

  final class DefaultImpls {
    public static void bar(Foo instance, String input) {
      System.out.println("I'm a " + instance + " saying " + input)
    }
  }
}
```

* Extract suffix for readability

* Fix incorrect sepearate for parameters in reflective descriptors

* Update interfaces test to be more exhaustive & cover different defaults

* Spotless

* Check for class existence first