Replace TODO_BLOCK stub with NOT_IMPLEMENTED
This replaces the TODO() placeholders in returned specs with a more correct (and less noisy in the IDE) `throw NotImplementedError("Stub!")`. I think this is a better approach to the code as it makes it clear it's just not implemented period, rather than possibly not implemented "yet".

No type token needed since it's a top-level type.