Deprecate mirror and element APIs
This deprecates existing mirror and element APIs, as they don't (and can't) return completely correct information about Kotlin types. This is just a first pass to soft-deprecate them and communicate that metadata should be used instead.

One thing not covered - `CodeBlock` accepts `TypeMirror` and `Element` as possible arguments. We could remove them, but it would be a behavior change. Thoughts?