Add basic support for default imports
This implements support for specifying custom default imports in `FileSpec`. The goal here is mostly targeting scripting generation, but I've also added a helper that allows for opt-in inclusion of kotlin's default imports to this (i.e. they would then be _excluded_ during writing).

Open to suggestions on other tests to add!