Avoid emitting public if shouldEmitPublicModifier returns false
Fixes #1341.

Previously we'd return true even if public was not in the implicit modifier list, which would cause the calling method to double emit Public.

Now we’ll explicitly filter public from the caller specified modifiers if shoudlEmitPublicModifier returns false.

I’ve added a couple tests that cover the failure and a couple that would have passed before this change but that increase coverage.