Do not emit superfluous `Unit` in return type position (#1559)
* Do not emit superfluous Unit in return type position

The only time it needs produced is when the function body is an expression in a non-ctor/getter/setter.

This contains a behavior change which explicitly makes the default return type Unit. This was somewhat the case before, except when you emitted a function body which could be turned into an expression. In this case you were able to get an implicit (i.e., unspecified) return type as none was declared. This would violate usages in explicit API contexts. A function with an omitted return type and a body which can be turned into an expression will now see an explicit Unit return type emitted. The fix is to declare the proper return type or remove the use of 'return'.

* Add change log entry in unreleased section