Append nested class names to alias during name lookup
Before this change, if name lookup for a ClassName located an
alias, we'd simply return that alias. This was broken in cases
when that ClassName had nested subclasses, as nested subclass
names would be discarded, instead of appended to the alias. With
this change, we're properly appending nested class names to the
alias.