Migrate TypeAliasSpec ad TypeVariableName to common.
Restored some of the functions in Util.kt that needed to wait for CodeBlock migration before they could be used.

Migrate several internal factory functions in TypeVariableName companion objects to the platform (TypeVariableName.jvm.kt).

There are some related but unmigrated types (such as CodeWriter), and this commit cannot be compiled at the moment.